日志模块基本完成

+++ jachin-blog/src/main/java/com/jachin/blog/mapper/MenuMapper.java
@@ -18,5 +18,5 @@ public interface MenuMapper extends BaseMapper<MenuEntity> {
      * @param uid user id
      * @return menus
      */
-    List<MenuEntity> listByUid(Integer uid);
+    List<MenuEntity> listByUid(Long uid);
 }
